+++ src/core/or/connection_edge.c
@@ -73,7 +73,7 @@
 #include "lib/crypt_ops/crypto_util.h"
 #include "feature/relay/dns.h"
 #include "feature/client/dnsserv.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
 #include "feature/dircache/dirserv.h"
 #include "feature/hibernate/hibernate.h"
 #include "feature/hs/hs_common.h"
