+++ src/feature/client/bridges.c
@@ -18,7 +18,9 @@
 #include "core/or/circuitbuild.h"
 #include "app/config/config.h"
 #include "core/mainloop/connection.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
+#include "feature/dirclient/dlstatus.h"
 #include "feature/client/entrynodes.h"
 #include "feature/nodelist/nodelist.h"
 #include "core/or/policies.h"
