+++ src/feature/control/control.c
@@ -56,7 +56,9 @@
 #include "feature/control/fmt_serverstatus.h"
 #include "lib/crypt_ops/crypto_rand.h"
 #include "lib/crypt_ops/crypto_util.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
+#include "feature/dirclient/dlstatus.h"
 #include "feature/dircache/dirserv.h"
 #include "feature/client/dnsserv.h"
 #include "feature/client/entrynodes.h"
