+++ src/feature/dirauth/dirvote.c
@@ -9,7 +9,8 @@
 #include "feature/dirauth/dircollate.h"
 #include "feature/dirauth/recommend_pkg.h"
 #include "feature/dirauth/voteflags.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
 #include "feature/dirauth/bwauth.h"
 #include "feature/dircache/dirserv.h"
 #include "feature/dirauth/guardfraction.h"
