+++ src/feature/dirauth/process_descs.c
@@ -19,7 +19,8 @@
 #include "core/or/policies.h"
 #include "feature/dirauth/keypin.h"
 #include "feature/dirauth/reachability.h"
-#include "feature/dircache/directory.h"
+#include "feature/dirclient/dlstatus.h"
+#include "feature/dircommon/directory.h"
 #include "feature/nodelist/networkstatus.h"
 #include "feature/nodelist/nodelist.h"
 #include "feature/nodelist/routerlist.h"
