+++ src/feature/hs/hs_client.c
@@ -19,7 +19,8 @@
 #include "lib/crypt_ops/crypto_format.h"
 #include "lib/crypt_ops/crypto_rand.h"
 #include "lib/crypt_ops/crypto_util.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
 #include "feature/hs/hs_cache.h"
 #include "feature/hs/hs_cell.h"
 #include "feature/hs/hs_circuit.h"
