+++ src/feature/nodelist/authcert.c
@@ -23,7 +23,9 @@
 #include "core/mainloop/mainloop.h"
 #include "core/or/policies.h"
 #include "feature/client/bridges.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
+#include "feature/dirclient/dlstatus.h"
 #include "feature/dircommon/fp_pair.h"
 #include "feature/nodelist/authcert.h"
 #include "feature/nodelist/dirlist.h"
