+++ src/feature/nodelist/dirlist.c
@@ -29,7 +29,7 @@
 #include "app/config/config.h"
 #include "core/or/policies.h"
 #include "feature/control/control.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
 #include "feature/nodelist/dirlist.h"
 #include "feature/nodelist/networkstatus.h"
 #include "feature/nodelist/nodelist.h"
