+++ src/feature/nodelist/networkstatus.c
@@ -51,7 +51,9 @@
 #include "feature/control/control.h"
 #include "lib/crypt_ops/crypto_rand.h"
 #include "lib/crypt_ops/crypto_util.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
+#include "feature/dirclient/dlstatus.h"
 #include "feature/dircache/dirserv.h"
 #include "feature/dirauth/reachability.h"
 #include "core/or/dos.h"
