+++ src/feature/nodelist/node_select.c
@@ -18,7 +18,8 @@
 #include "core/or/policies.h"
 #include "core/or/reasons.h"
 #include "feature/client/entrynodes.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
 #include "feature/nodelist/dirlist.h"
 #include "feature/nodelist/microdesc.h"
 #include "feature/nodelist/networkstatus.h"
