+++ src/feature/nodelist/routerlist.c
@@ -69,8 +69,10 @@
 #include "feature/client/bridges.h"
 #include "feature/control/control.h"
 #include "feature/dirauth/mode.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
 #include "feature/dircache/dirserv.h"
+#include "feature/dirclient/dlstatus.h"
 #include "feature/dirauth/reachability.h"
 #include "feature/dirauth/process_descs.h"
 #include "feature/nodelist/authcert.h"
