+++ src/feature/relay/router.c
@@ -16,7 +16,8 @@
 #include "lib/crypt_ops/crypto_rand.h"
 #include "lib/crypt_ops/crypto_util.h"
 #include "lib/crypt_ops/crypto_curve25519.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
 #include "feature/dircache/dirserv.h"
 #include "feature/dirauth/process_descs.h"
 #include "feature/relay/dns.h"
