+++ src/feature/rend/rendservice.c
@@ -19,7 +19,8 @@
 #include "lib/crypt_ops/crypto_dh.h"
 #include "lib/crypt_ops/crypto_rand.h"
 #include "lib/crypt_ops/crypto_util.h"
-#include "feature/dircache/directory.h"
+#include "feature/dircommon/directory.h"
+#include "feature/dirclient/dirclient.h"
 #include "feature/hs/hs_common.h"
 #include "feature/hs/hs_config.h"
 #include "core/mainloop/mainloop.h"
